% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{api.setup}
\alias{api.setup}
\title{Set up Onboard API keys and URL in system environment}
\usage{
api.setup(api_type = "prod")
}
\arguments{
\item{api_type}{Provide the API client name.}
}
\value{
No return value, sets API url and API key in the system environment.
}
\description{
Set the Onboard API URL and API keys in the system environment.
}
