% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_buildings}
\alias{get_buildings}
\title{Buildings}
\usage{
get_buildings(id)
}
\arguments{
\item{id}{(Optional) An integer if you want information on a particular entity. Returns all entities unless this argument is provided.}
}
\value{
A data.frame of all building information.
}
\description{
Retrieve buildings that you have access to.
}
