% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Time-Series Data}
\usage{
get_timeseries(start_time, end_time, point_ids)
}
\arguments{
\item{start_time}{Start Time in UTC.}

\item{end_time}{End Time in UTC.}

\item{point_ids}{Point IDs for which timeseries data needs to be queried.}
}
\value{
A wide data.frame of time-series data, with timestamp and all requested point IDs as columns.
}
\description{
Provides clean time-series
}
