% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries_raw}
\alias{get_timeseries_raw}
\title{Raw Time-Series Data}
\usage{
get_timeseries_raw(start_time, end_time, point_ids)
}
\arguments{
\item{start_time}{Start Time in UTC.}

\item{end_time}{End Time in UTC.}

\item{point_ids}{Point IDs for which timeseries data needs to be queried.}
}
\value{
A long data.frame of time series data, with point id, timestamp, and raw point values as columns.
}
\description{
Retrieves timeseries data in raw format.
}
