% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_users}
\alias{get_users}
\title{Users}
\usage{
get_users(id)
}
\arguments{
\item{id}{(Optional) An integer if you want information on a particular entity. Returns all entities unless this argument is provided.}
}
\value{
A data.frame of all user information.
}
\description{
Retrieve all user info in your organization.
}
