% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_selector.R
\name{PointSelector}
\alias{PointSelector}
\title{PointSelector}
\usage{
PointSelector()
}
\value{
An empty named list of possible point selection criteria.
}
\description{
A list of parameters to query metadata.
}
\examples{
\dontrun{
query <- PointSelector()

query$buildings <- 101
query$equipment_types <- 'ahu'
query$point_types <- c('Supply Air Temperature','Supply Air Static Pressure')
}

}
