% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_selector.R
\name{select_points}
\alias{select_points}
\title{Select Points}
\usage{
select_points(query)
}
\arguments{
\item{query}{query supplied from PointSelector.}
}
\value{
A named list of all the points requested by the query.
}
\description{
Returns a list of ids based on the input query from PointSelector. Uses http POST call to query data.
}
\examples{
\dontrun{
query <- PointSelector()

query$buildings <- 427
query$equipment_types <- 'ahu'
query$point_types <- c('Supply Air Temperature','Supply Air Static Pressure')

selection <- select_points(query)
}

}
