% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{upload_staging}
\alias{upload_staging}
\title{Upload to Staging Area}
\usage{
upload_staging(building, data_to_upload, skip_topics = FALSE, verbose = TRUE)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry is considered.}

\item{data_to_upload}{A data.frame to upload to the staging area. Must contain e.equip_id and p.topic columns.}

\item{skip_topics}{Logical. If True, the uploaded topics will be assigned `__SKIP__` equip_id.}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
A named list containing any errors that may have occured during data upload.
}
\description{
Uploads data to the staging area.
}
